requirements.txt
openai
elevenlabs
python-dotenv
PyYAML
requests
google-cloud-storage